+++ inst/make_ppk_models.R
@@ -109,8 +109,53 @@ mod_amikacin_2cpt_Burdet2015 <- list(
   covariates  = c("CLCREAT4H","TBW","PoverF"),
   sigma       = c(additive_a = 0.2, proportional_b = 0.1))
 
+mod_piperacillin_2cpt_Roberts2010 <- list(
+  ppk_model   = RxODE::RxODE({
+    centr(0)   = 0;
+    # Time lag from dose infuser to patient
+    TVLAGTIME  = THETA_LAGTIME
+    LAGTIME    = TVLAGTIME*exp(ETA_LAGTIME)
+    # -------------------------------------
+    TVCl  = THETA_Cl*(TBW/70);
+    TVV1  = THETA_V1;
+    TVV2  = THETA_V2;
+    TVQ   = THETA_Q;
+    Cl    = TVCl*exp(ETA_Cl + KAPPA_Cl);
+    Cli   = TVCl*exp(ETA_Cl);
+    V1    = TVV1*exp(ETA_V1 + KAPPA_V1);
+    V1i   = TVV1*exp(ETA_V1);
+    V2    = TVV2*exp(ETA_V2);
+    Q     = TVQ *exp(ETA_Q);
+    ke    = Cl/V1;
+    k12   = Q/V1;
+    k21   = Q/V2;
+    Cc    = centr/V1;
+    d/dt(centr)  = - ke*centr - k12*centr + k21*periph;
+    d/dt(periph) =            + k12*centr - k21*periph;
+    lag(centr)   =   LAGTIME;
+    d/dt(AUC)    =   Cc;
+  }),
+  error_model = function(f,sigma){
+    g <- sigma[1] + sigma[2]*f
+    return(g)
+  },
+  theta = c(THETA_LAGTIME=0.07,THETA_Cl=17.1,THETA_V1=7.2,THETA_V2=17.8,
+            THETA_Q=52.0),
+  omega = lotri::lotri({ETA_LAGTIME + ETA_Cl + ETA_V1 + ETA_V2 + ETA_Q ~
+      c(0.1747673,
+        0.00     ,  0.08507985,
+        0.00     ,  0.00      ,  0.0673745,
+        0.00     ,  0.00      ,  0.00     ,  0.429067,
+        0.00     ,  0.00      ,  0.00     ,  0.00    ,  0.2247455)}),
+  pi_matrix = lotri::lotri({KAPPA_Cl + KAPPA_V1 ~
+      c(0.1934626,
+        0.00     ,  0.05783106)}),
+  covariates  = c("TBW"),
+  sigma       = c(additive_a = 3.2, proportional_b = 0.253))
+
 ## save the models in a .rda data file
 save(mod_tobramycin_2cpt_fictional,
      mod_vancomycin_2cpt_Goti2018,
      mod_amikacin_2cpt_Burdet2015,
+     mod_piperacillin_2cpt_Roberts2010,
      file="data/lib_ppk_model.rda")