+++ tests/testthat/test_classoutput.R
@@ -68,3 +68,72 @@ test_that("poso_estim_sir returns the expected objects", {
   expect_equal("rxSolve" %in% attributes(p01_sir_nomod$model)$class,
                FALSE)
 })
+
+df_patient06_pipera <- data.frame(ID=6,TIME=c(0.0,8,16,24,32,40,47.9,48,
+                                         56,64,71.9,72,80,88,95.9,
+                                         96,104,112,120),
+                                  DV=c(NA,NA,NA,NA,NA,NA,63.0,NA,NA,
+                                       NA,19.7,NA,NA,NA,31,NA,NA,NA,16),
+                                  AMT=c(4000,4000,4000,4000,4000,4000,NA,
+                                        4000,4000,4000,NA,4000,4000,4000,
+                                        NA,4000,4000,4000,NA),
+                                  DUR=c(8,8,8,8,8,8,NA,8,8,8,NA,8,8,8,NA,
+                                        8,8,8,NA),
+                                  EVID=c(1,1,1,1,1,1,0,1,1,1,0,1,1,1,0,
+                                         1,1,1,0),TBW=92,
+                                  OCC=c(1,1,1,1,1,1,1,2,2,2,2,3,3,3,3,
+                                        4,4,4,4))
+
+patient06_pipera <- posologyr(prior_model=mod_piperacillin_2cpt_Roberts2010,
+                              dat=df_patient06_pipera)
+
+test_that("poso_simu_pop returns the expected objects (with IOV)", {
+  p06_pop_mod    <- poso_simu_pop(patient06_pipera)
+  p06_pop_nomod  <- poso_simu_pop(patient06_pipera,return_model = FALSE)
+
+  expect_equal("data.frame" %in% attributes(p06_pop_mod$eta)$class,
+               TRUE)
+  expect_equal("rxSolve" %in% attributes(p06_pop_mod$model)$class,
+               TRUE)
+  expect_equal("data.frame" %in% attributes(p06_pop_nomod$eta)$class,
+               TRUE)
+  expect_equal("rxSolve" %in% attributes(p06_pop_nomod$model)$class,
+               FALSE)
+})
+
+test_that("poso_estim_map returns the expected objects (with IOV)", {
+  p06_map_mod    <- poso_estim_map(patient06_pipera)
+  p06_map_nomod  <- poso_estim_map(patient06_pipera,return_model = FALSE)
+
+  expect_equal("data.frame" %in% attributes(p06_map_mod$eta)$class,
+               FALSE)
+  expect_equal(attributes(p06_map_mod$eta)$class,
+               NULL)
+  expect_equal("rxSolve" %in% attributes(p06_map_mod$model)$class,
+               TRUE)
+  expect_equal("data.frame" %in% attributes(p06_map_nomod$eta)$class,
+               FALSE)
+  expect_equal(attributes(p06_map_nomod$eta)$class,
+               NULL)
+  expect_equal("rxSolve" %in% attributes(p06_map_nomod$model)$class,
+               FALSE)
+})
+
+test_that("poso_estim_sir returns the expected objects (with IOV)", {
+  p06_sir_mod   <- poso_estim_sir(patient06_pipera,
+                                  n_sample=5e2,
+                                  n_resample=1e2)
+  p06_sir_nomod <- poso_estim_sir(patient06_pipera,
+                                  n_sample=5e2,
+                                  n_resample=1e2,
+                                  return_model = FALSE)
+
+  expect_equal("data.frame" %in% attributes(p06_sir_mod$eta)$class,
+               TRUE)
+  expect_equal("rxSolve" %in% attributes(p06_sir_mod$model)$class,
+               TRUE)
+  expect_equal("data.frame" %in% attributes(p06_sir_nomod$eta)$class,
+               TRUE)
+  expect_equal("rxSolve" %in% attributes(p06_sir_nomod$model)$class,
+               FALSE)
+})