+++ NEWS.md
@@ -1,14 +1,18 @@
 # posologyr v0.1.1.9000
 
+* `poso_estim_sir()` estimates the posterior distribution of individual 
+parameters by Sequential Importance Resampling (SIR). It is roughly 25 times 
+faster than `poso_estim_mcmc()` for 1000 samples.
 * `poso_estim_map()` allows the estimation of the individual parameters by 
 adaptive MAP forecasting (cf. doi: 10.1007/s11095-020-02908-7) with 
 `adapt=TRUE`.
-* `poso_estim_sir()` allows the estimation of the posterior distribution 
-of individual parameters by Sequential Importance Resampling (SIR). It is 
-roughly 25 times faster than `poso_estim_mcmc()` for 1000 samples.
+* `poso_simu_pop()`, `poso_estim_map()`, and `poso_estim_sir()` now support 
+models with both inter-individual (IIV) and inter-occasion variability (IOV).
 * `MASS:mvrnorm` is replaced by `mvtnorm::rmvnorm` for multivariate normal 
 distributions. 
 * Input validation is added to all exported functions.
+* `poso_estim_map()` now uses method="L-BFGS-B" in optim for better convergence 
+of the algorithm.
 * `poso_inter_cmin()` now uses method="L-BFGS-B" in optim for better convergence 
 of the algorithm.
 * `poso_dose_conc()` is the new name of `poso_dose_ctime()`.